+++ belt/models.py
@@ -1,13 +1,19 @@
+import io
 import warnings
+from typing import Dict, List, Optional, Tuple
 
 from django.core.exceptions import ValidationError
+from django.core.files.base import ContentFile
 from django.db import models
 from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
+
+from belt.files import UploadToDir
 
 
 class LogicDeleteMixin(models.Model):
     """Mixin to handle basic functionality of a logic delete"""
+
     deleted = models.BooleanField(default=False)
     date_deleted = models.DateTimeField(null=True, blank=True)
 
@@ -63,17 +69,17 @@ def transition_handler_decorator(func):
 class StatusMixin(models.Model):
     """Mixin to handle status changes"""
 
-    STATUS_FIELD = "status"
-    ALLOWED_TRANSITIONS = []
-    FORBIDDEN_TRANSITIONS = []
-    TRANSITION_HANDLERS = {}
+    STATUS_FIELD: str = "status"
+    ALLOWED_TRANSITIONS: List[Tuple[str, str]] = []
+    FORBIDDEN_TRANSITIONS: List[Tuple[str, str]] = []
+    TRANSITION_HANDLERS: Dict[Tuple[str, str], str] = {}
 
     class Meta:
         abstract = True
 
     def __init__(self, *args, **kwargs):
         """Init _handling_transition value to False."""
-        self._handling_transition = False
+        self._handling_transition: bool = False
         super().__init__(*args, **kwargs)
         self._original_status = getattr(self, self.STATUS_FIELD)
 
@@ -82,15 +88,19 @@ class StatusMixin(models.Model):
         if hasattr(self, "_transition"):
             delattr(self, "_transition")
 
-    def get_status_transition(self):
+    def get_status_transition(self) -> Optional[Tuple[str, str]]:
         """Get status transition."""
         if self.pk:
             if hasattr(self, "_transition"):
                 return self._transition
             previous = self._meta.model.objects.get(pk=self.pk)
             if previous.status != getattr(self, self.STATUS_FIELD):
-                self._transition = previous.status, getattr(self, self.STATUS_FIELD)
+                self._transition: Tuple[str, str] = (
+                    previous.status,
+                    getattr(self, self.STATUS_FIELD),
+                )
                 return self._transition
+        return None
 
     def validate_transition(self):
         """Validates the transition."""
@@ -120,11 +130,11 @@ class StatusMixin(models.Model):
                 )
             )
 
-    def pre_status_handler(self, transition):
+    def pre_status_handler(self, transition: Tuple[str, str]):
         """Method used to execute code before the status handler is called."""
         pass
 
-    def post_status_handler(self, transition):
+    def post_status_handler(self, transition: Tuple[str, str]):
         """Method used to execute code after the status handler is called."""
         pass
 
@@ -162,3 +172,133 @@ class StatusMixin(models.Model):
             self.pre_status_handler(transition)
             transition_handler(self)
             self.post_status_handler(transition)
+
+
+class ExporterModel(models.Model):
+
+    IDLE, RUNNING, FINISHED, ERROR = 0, 1, 2, 3
+    STATUSES = (
+        (IDLE, _("Idle")),
+        (RUNNING, _("Running")),
+        (FINISHED, _("Finished")),
+        (ERROR, _("Error")),
+    )
+
+    data = models.FileField(
+        _("data"),
+        upload_to=UploadToDir("export", random_name=True),
+        null=True,
+        blank=True,
+    )
+
+    status = models.SmallIntegerField(
+        _("status"), choices=STATUSES, default=IDLE, blank=True
+    )
+    items = models.PositiveIntegerField(_("items"), null=True, blank=True)
+
+    class Meta:
+        abstract = True
+
+    def __str__(self):
+        return f"Exporter #{self.pk}"
+
+    def get_queryset(self):
+        """Gets the QuerySet to export."""
+        raise NotImplementedError()
+
+    def exporter(self):
+        """Gets the import helper function that does the export."""
+        raise NotImplementedError()
+
+    def exporter_task(self):
+        """Gets the async task that does the exporter."""
+        raise NotImplementedError()
+
+    def export_data(self, async_process=False):
+        """Executes the import process."""
+        if self.status != self.IDLE:
+            return  # Only import if is idle
+        if async_process:
+            task = self.exporter_task()
+            task.delay(self.pk)
+        else:
+            self.items = 0
+            self.status = self.RUNNING
+            self.save()
+            # Export data and save it in the file field
+            exporter = self.exporter()
+            output = io.StringIO()
+            items = exporter(self.get_queryset(), output)
+            output.seek(0)
+            content = output.getvalue()
+            self.data.save("export.csv", content=ContentFile(content.encode("utf-8")))
+            self.items = items
+            self.status = self.FINISHED
+            self.save()
+
+
+class ImporterModel(models.Model):
+    """Abstract model to implement an asynchronous import process."""
+
+    IDLE, RUNNING, FINISHED, ERROR = 0, 1, 2, 3
+    STATUSES = (
+        (IDLE, _("Idle")),
+        (RUNNING, _("Running")),
+        (FINISHED, _("Finished")),
+        (ERROR, _("Error")),
+    )
+
+    data = models.FileField(
+        _("data"), upload_to=UploadToDir("imports", random_name=False)
+    )
+
+    status = models.SmallIntegerField(
+        _("status"), choices=STATUSES, default=IDLE, blank=True
+    )
+    items = models.PositiveIntegerField(_("items"), null=True, blank=True)
+
+    class Meta:
+        abstract = True
+
+    def __str__(self):
+        return f"Importer #{self.pk}"
+
+    def source(self):
+        """Creates the source for the importer using the data file."""
+        file_content = self.data.read()
+        source = io.BytesIO()
+        source.write(file_content)
+        source.seek(0)
+        return source
+
+    def importer(self):
+        """Gets the import helper function that does the import."""
+        raise NotImplementedError()
+
+    def importer_task(self):
+        """Gets the async task that does the import."""
+        raise NotImplementedError()
+
+    def import_data(self, async_process=False):
+        """Executes the import process."""
+        if self.status != self.IDLE:
+            return  # Only import if is idle
+        if async_process:
+            task = self.importer_task()
+            task.delay(self.pk)
+        else:
+            self.items = 0
+            self.status = self.RUNNING
+            self.save()
+            # Import line by line to don't load all the file in memory
+            self.data.seek(0)
+            importer = self.importer()
+            try:
+                items = importer(source=self.source())
+                self.items = items.count()
+                self.status = self.FINISHED
+                self.save()
+            except Exception as exception:
+                self.status = self.ERROR
+                self.save()
+                raise exception