+++ belt/rest_framework/mixins.py
@@ -1,6 +1,8 @@
-from rest_framework.serializers import Serializer
+from typing import Dict, List, Type, Union
+
 from django.db.models import QuerySet
 from rest_framework.response import Response
+from rest_framework.serializers import Serializer
 
 
 class ActionSerializersMixin:
@@ -9,7 +11,7 @@ class ActionSerializersMixin:
     each action.
     """
 
-    action_serializers = {}
+    action_serializers: Dict[str, Type[Serializer]] = {}
 
     def get_serializer_class(self):
         if hasattr(self, "action_serializers"):
@@ -23,8 +25,8 @@ class SimpleSerializationMixin:
     Mixin to simplify serialization of object instances.
     """
 
-    def serialize_instance(self, data) -> Serializer:
-        serializer_class = self.get_serializer_class()
+    def serialize_instance(self, data) -> Union[Dict, List]:
+        serializer_class = self.get_serializer_class()  # type: ignore
         serializer = serializer_class(instance=data)
 
         return serializer.data
@@ -36,12 +38,12 @@ class SimplePaginationMixin:
     """
 
     def paginate(self, queryset: QuerySet) -> Response:
-        page = self.paginate_queryset(queryset)
+        page = self.paginate_queryset(queryset)  # type: ignore
         if page is not None:
-            serializer = self.get_serializer(page, many=True)
-            return self.get_paginated_response(serializer.data)
+            serializer = self.get_serializer(page, many=True)  # type: ignore
+            return self.get_paginated_response(serializer.data)  # type: ignore
 
-        serializer = self.get_serializer(queryset, many=True)
+        serializer = self.get_serializer(queryset, many=True)  # type: ignore
         return Response(serializer.data)
 
 
@@ -51,6 +53,6 @@ class SimpleValidationMixin:
     """
 
     def validate(self, data: dict, raise_exception=True) -> Serializer:
-        serializer = self.get_serializer(data=data)
+        serializer = self.get_serializer(data=data)  # type: ignore
         serializer.is_valid(raise_exception=raise_exception)
         return serializer