+++ tests/urls.py
@@ -1,1 +1,3 @@
-urlpatterns = []
+from typing import List
+
+urlpatterns: List = []